give up on rust_codegen, just dump types to stdout
