+++ Cargo.toml
@@ -15,8 +15,32 @@ include = ["/rust/*.rs", "/Cargo.toml", "/schemas/**", "/topics/**", "/README.md
 [lib]
 path = "rust/lib.rs"
 
+[[bin]]
+# register build.rs as binary so people can view generated code directly as well
+# https://stackoverflow.com/a/53728256/1544347
+name = "generate-schema-types"
+path = "rust/build.rs"
+required-features = ["build_deps"]
+
+[features]
+build_deps = ["prettyplease", "schemars", "syn", "typify"]
+
 [dependencies]
 serde = { version = "1.0", features = ["derive"]}
 serde_json = "1.0"
 serde_yaml = "0.9"
-regress = "0.5.0"
+regress = "0.5.0"  # required by generated code
+
+# redeclare all build-dependencies here so that the build.rs can be used
+# standalone to view types
+prettyplease = { version = "0.2.4", optional = true }
+schemars = { version = "0.8.12", optional = true }
+syn = { version = "2.0.11", optional = true }
+typify = { version = "0.0.12-dev", git = "https://github.com/oxidecomputer/typify", optional = true }
+
+[build-dependencies]
+prettyplease = "0.2.4"
+schemars = "0.8.12"
+serde_json = "1.0.95"
+syn = "2.0.11"
+typify = { version = "0.0.12-dev", git = "https://github.com/oxidecomputer/typify" }
