+++ rust/build.rs
@@ -1,14 +1,33 @@
-use std::io::Write;
+use std::path::Path;
+
+use typify::{TypeSpace, TypeSpaceSettings};
+
+fn generate_schema(schema_path: &str, output_module: &str) -> String {
+    println!("cargo:rerun-if-changed={schema_path}");
+    let json = std::fs::read_to_string(schema_path).expect("Read schema JSON file");
+
+    let schema = serde_json::from_str::<schemars::schema::RootSchema>(&json).unwrap();
+
+    let mut type_space = TypeSpace::new(TypeSpaceSettings::default().with_struct_builder(true));
+    type_space.add_root_schema(schema).unwrap();
+
+    let code = prettyplease::unparse(&syn::parse2::<syn::File>(type_space.to_stream()).unwrap());
+
+    format!("pub mod {output_module} {{
+    use serde::{{Deserialize, Serialize}};
+    {code}
+    }}")
+}
 
 fn main() {
-    if !std::path::Path::new("rust_codegen/Cargo.toml").is_file() {
-        return;
+    let mut module_code = String::new();
+    module_code.push_str(
+        &generate_schema("schemas/ingest-metrics.v1.schema.json", "ingest_metrics_v1")
+    );
+
+    if let Ok(target_dir) = std::env::var("OUT_DIR") {
+        std::fs::write(Path::new(&target_dir).join("schema_types.rs"), module_code).expect("Failed to write schema_types.rs");
+    } else {
+        println!("{module_code}");
     }
-
-    let output = std::process::Command::new("cargo")
-            .args(["run", "--manifest-path", "rust_codegen/Cargo.toml"])
-            .output()
-            .unwrap();
-
-    std::io::stdout().write(&output.stdout).unwrap();
 }
