--- rust_codegen/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "sentry-kafka-schemas-codegen"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-prettyplease = "0.2.4"
-schemars = "0.8.12"
-serde_json = "1.0.95"
-syn = "2.0.11"
-typify = { version = "0.0.12-dev", git = "https://github.com/oxidecomputer/typify" }
